+++ src/robopilot/refiner/__init__.py
@@ -1,0 +1,5 @@
+"""ProjectSpec refinement helpers."""
+
+from robopilot.refiner.spec_refiner import refine_spec
+
+__all__ = ["refine_spec"]